use mysql pool connection for session, separate config by APP_STATE env variable
