+++ server/app/controllers/passport.js
@@ -1,7 +1,7 @@
 // load up the instagram strategy
 var InstagramStrategy = require('passport-instagram').Strategy;
 
-var config = require('../../config');
+var config = require('../../config_' + process.env.APP_STATE);
 var usersController = require('./users');
 var UserDA = require('../data_access/user');
 var mediaTypeMapper = require('../models/media_type_mapper');
